Combined
First try to combine, things missing, working on them
Now have a common dataset : easyshare and life, then merged to job, then taking only last wave for each id. Tried to do most of the things on this whole dataset. Other one much smaller with deleted irrelevant columns + dropped of variables w/ NA > 2000. Some descriptives, the grouped FS of Lino but on the right dataset and the whole dataset (not a sample of it), some parts of Henry, not mixed anymore.